agents/school_analytics_agent_david/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-openai==0.2.8
google-cloud-firestore>=2.19.0
httpx>=0.26.0
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
statsmodels>=0.14.0